pyproject.toml
[tool.poetry]
name = "talebi-service"
version = "0.1.0"
description = ""
authors = ["Hamed Feizabadi <hamedfz19@gmail.com>"]
readme = "README.md"
packages = [{include = "talebi_service"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
Pillow = "^9.4.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.26.0"}
faker = "^18.3.2"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
django-filter = "^23.1"
whitenoise = "^6.4.0"
requests = "^2.29.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
